+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added Go agent
 -   Added new compiler class `GoCompiler`
 -   Added new stager type `multi_go_exe`
+-   Added autorun capability for agents on checkin through Starkiller menu
 -   Added go to install script
 
 
@@ -96,6 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Fixed
+-   Fixed Powershell agent overwritting results for C# taskings
 
 ### Security